[dev] guest user's cart - Unauthenticated users can now add items to their cart and while on the cart page, you can see the cart update

+++ django_ecom_app/app_ecommerce/store/store/models.py
@@ -14,7 +14,7 @@ class Customer(models.Model):
 # Product Model
 class Product(models.Model):
     name = models.CharField(max_length=200, null=True)
-    price = models.FloatField()
+    price = models.DecimalField(max_digits=7, decimal_places=2)
     # 'digital' field tells us essentially if the product needs to be shipped 
     digital = models.BooleanField(default=False, null=True, blank=True)
     image = models.ImageField(null=True, blank=True)
